+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [NEXT] - 
+
+## [14.6.0] - 2025-04-24
+
+### Changed:
+
+- Updated underlying Onfido native SDK version:
+  - iOS 32.5.x (up from 32.4.x)
+  - Android 22.7.x (up from 22.6.x)
+
+### Fixed:
+
+- Allowed `captureDocument.allowedDocumentTypes` typing to have multiple document types
+- Changed `flowSteps` and `Theme` to optional to be consistent with documentation
+
 ## [14.5.0] - 2025-04-08
 
 ### Changed: